TyphoonApi/requirements.txt
astroid==1.6.3
autopep8==1.3.5
certifi==2018.1.18
chardet==3.0.4
Django==2.0.4
django-filter==1.1.0
django-tables2==1.21.1
djangorestframework==3.8.2
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.2.1
Markdown==2.6.11
mccabe==0.6.1
Pillow==5.1.0
pycodestyle==2.4.0
pylint==1.8.4
pytz==2018.3
reportlab==3.4.0
requests==2.18.4
six==1.11.0
style==1.1.0
update==0.0.1
urllib3==1.22
wrapt==1.10.11
